添加json11支持

+++ include/reply.hpp
@@ -32,7 +32,7 @@ public:
 
         // reply fixed header
         headers[0].name = "Content-Length";
-        headers[0].value = std::to_string(content.size());
+        headers[0].value = std::to_string((long long unsigned)content.size());
         headers[1].name = "Content-Type";
         headers[1].value = "text/html";
         headers[2].name = "Connection";
